requirements.txt
# Backend Web
flask==3.0.3
python-dotenv==1.0.1

# Spotify API
spotipy==2.24.0

# Machine Learning
scikit-learn==1.5.2
xgboost==2.1.1
pandas==2.2.3
numpy==1.26.4
shap==0.46.0

# NLP y Análisis de Letras
pysentimiento==0.7.3

# Web Scraping (letras)
requests==2.32.3
beautifulsoup4==4.12.3
unidecode==1.3.8

# Utilidades
tqdm==4.66.5
joblib==1.4.2

# Producción
gunicorn==23.0.0
